requirements.txt
fastapi
uvicorn
ollama
sentence-transformers
faiss-cpu
numpy
requests
beautifulsoup4
lxml
ddgs
jinja2
pydantic